fix(pg): merge GPU utils into mooncake_backend.h, fix reduceKernel for mt_bfloat16
- Move GPU utility functions (getCurrentGPUStream, getGPUStreamFromPool,
  currentGPUDevice, kGPUDevice, kGPUDeviceType) from standalone
  mooncake_pg_gpu_utils.h into mooncake_backend.h, which both
  mooncake_worker_host.cpp and mooncake_backend.cpp already include.
- Delete mooncake_pg_gpu_utils.h.
- In reduceKernel, add #ifdef __MUSA__ path using float accumulator +
  fminf/fmaxf because mt_bfloat16 lacks arithmetic operators (+=, *=,
  std::min, std::max). CUDA path unchanged.
